Remove dependency on JSON gem (#14)
* remove dependency on JSON gem

See http://www.mikeperham.com/2016/02/09/kill-your-dependencies/ ...
Ruby knows how to parse JSON since 1.9 and if anybody is still using the
unsupported 1.8 they should be requiring this dependency themselves.

* Add a .gitignore file and remove Gemfile.lock

This adds the `.gitignore` file that would have been generated by
`bundle gem`. It also removes the `Gemfile.lock` file since [it wouldn't
be honored anyways, when the gem is
installed](http://yehudakatz.com/2010/12/16/clarifying-the-roles-of-the-gemspec-and-gemfile/).